+++ demo/feature_test.py
@@ -19,7 +19,15 @@ rb.Game.debug = True
 
 main = rb.Scene()
 
-test = rb.GameObject({"pos": rb.Vector(300, 300)}).add(rb.Rectangle({}))
+test = rb.GameObject({
+    "pos": rb.Vector(300, 300)
+}).add(rb.Rectangle({
+    "width": 50,
+    "height": 100,
+    "color": rb.Color.red
+})).add(rb.Slider({
+    "slider_origin_offset": rb.Vector(200, -50),
+}))
 
 print(test.get(rb.Rectangle).radius)