+++ rubato/utils/__init__.py
@@ -16,3 +16,4 @@ from .radio import Radio
 from .rb_input import Input
 from .sound import Sound
 from .quickselect import quickselect
+from .debug import Debug